BT sync: decimal-millisecond trims + the fine-adjustment ruler
T1: widen the per-device Bluetooth SYNC trim from Int to Double
milliseconds end to end and add BTSyncTrim.quantise, which snaps to
0.1 ms without binary-float dust. SyncTiming.totalDelayNanos widens with
it — a precision gain for the local sink, which still passes whole ms.
The on-disk store keeps schemaVersion 1 on purpose: a v1 file's bare
JSON integers decode into Double unchanged.

T4: BTSyncRulerView — a ticked tape sliding under a fixed centre
pointer, deliberately not a thumb on a track so it cannot be mistaken
for a volume slider. Drag gearing runs from 0.1 ms per ~3 pt up to
2 ms/pt by drag speed, with the slow gearing forced while option is
held. The pointer uses the gold token, which already resolves through
the Appearance dial, so it follows the system highlight for anyone who
picked that. Full keyboard and VoiceOver surface.

Fixes a latent crash found while building it: reading NSApp.currentEvent
force-unwraps, and NSApp is nil in a narrowly filtered test run. The
same pattern still exists in DeviceRowView and is folded into T6.

Plan: docs/plans/PLAN-BT-SYNC-DRAWER.md

+++ AudiouterCore/Sources/AudiouterCore/BTSyncedSink.swift
@@ -58,7 +58,7 @@ enum BTReferenceTimeline {
         presentationDelayMs: Int,
         btOnlyBufferMs: Int,
         deviceOffsetMs: Int,
-        trimMs: Int
+        trimMs: Double
     ) -> Int64 {
         SyncTiming.totalDelayNanos(
             presentationDelayMs: composition.airPlayPresent ? presentationDelayMs : btOnlyBufferMs,
@@ -915,7 +915,7 @@ final class BTSyncedSink: @unchecked Sendable {
     private var sinksByUID: [String: BTDeviceSink] = [:]
     private var composition = BTGroupComposition(airPlayPresent: false, macLocalPresent: false)
     private var offsetMsByUID: [String: Int] = [:]
-    private var trimMsByUID: [String: Int] = [:]
+    private var trimMsByUID: [String: Double] = [:]
     private var desiredRunning = false
 
     init(
@@ -1004,7 +1004,7 @@ final class BTSyncedSink: @unchecked Sendable {
     }
 
     /// The settable per-device signed manual trim (ms).
-    func setTrimMs(_ ms: Int, forDeviceUID uid: String) {
+    func setTrimMs(_ ms: Double, forDeviceUID uid: String) {
         let sink = tableLock.withLock { () -> BTDeviceSink? in
             guard trimMsByUID[uid] != ms else { return nil }
             trimMsByUID[uid] = ms
